llm/requirements_kube.txt
torch-model-archiver==0.8.1
kubernetes==28.1.0
kserve==0.11